pyproject.toml
[project]
name = "billy"
version = "0.1.0"
description = "AI-powered legislation understanding"
requires-python = ">=3.13"
dependencies = [
    "fastapi<1.0.0",
    "google-genai<2.0.0",
    "pydantic<3.0.0",
    "requests<3.0.0",
    "uvicorn<1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff==0.12.2",
    "black==23.11.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 120
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.git
  | \__pycache__
  | \venv
  | build
  | dist
  | .ruff_cache
)/
'''

[tool.ruff]
line-length = 120
target-version = "py313"
select = ["E", "F", "I"]
exclude = [
    ".git",
    ".ruff_cache",
    "build",
    "dist",
    "venv",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
